discovery-plugin/requirements.txt
APScheduler==3.10.0
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.0.1
confluent-kafka==2.0.2
idna==3.4
kafka-python==2.0.2
#pytz==2022.7.1
pytz
pytz-deprecation-shim==0.1.0.post0
rec-avro==0.0.4
requests==2.28.2
six==1.16.0
tinaa-utils
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14

